requirements.txt
accelerate==1.12.0
peft==0.18.1
bitsandbytes==0.49.1
transformers==4.57.6
trl==0.27.0
datasets==4.5.0
python-dotenv==1.2.1
tensorboardX==2.6.4
levenshtein==0.27.3